+++ Runtime/Scripts/Beacon/BeaconConnection.jslib
