+++ Runtime/Scripts/Beacon/IBeaconConnector.cs.meta
