+++ Runtime/Scripts/Helpers/UnityMainThreadDispatcher.cs
@@ -11,135 +11,138 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-using UnityEngine;
+using System;
 using System.Collections;
 using System.Collections.Generic;
-using System;
 using System.Threading.Tasks;
+using UnityEngine;
 
-/// Author: Pim de Witte (pimdewitte.com) and contributors, https://github.com/PimDeWitte/UnityMainThreadDispatcher
-/// <summary>
-/// A thread-safe class which holds a queue with actions to execute on the next Update() method. It can be used to make calls to the main thread for
-/// things such as UI Manipulation in Unity. It was developed for use in combination with the Firebase Unity plugin, which uses separate threads for event handling
-/// </summary>
-public class UnityMainThreadDispatcher : MonoBehaviour
+namespace TezosSDK.Helpers
 {
-    private static readonly Queue<Action> _executionQueue = new();
-
-    public void Update()
+    /// Author: Pim de Witte (pimdewitte.com) and contributors, https://github.com/PimDeWitte/UnityMainThreadDispatcher
+    /// <summary>
+    /// A thread-safe class which holds a queue with actions to execute on the next Update() method. It can be used to make calls to the main thread for
+    /// things such as UI Manipulation in Unity. It was developed for use in combination with the Firebase Unity plugin, which uses separate threads for event handling
+    /// </summary>
+    public class UnityMainThreadDispatcher : MonoBehaviour
     {
-        lock (_executionQueue)
+        private static readonly Queue<Action> _executionQueue = new();
+
+        public void Update()
         {
-            while (_executionQueue.Count > 0)
+            lock (_executionQueue)
             {
-                _executionQueue.Dequeue().Invoke();
-            }
-        }
-    }
-
-    /// <summary>
-    /// Locks the queue and adds the IEnumerator to the queue
-    /// </summary>
-    /// <param name="action">IEnumerator function that will be executed from the main thread.</param>
-    public void Enqueue(IEnumerator action)
-    {
-        lock (_executionQueue)
-        {
-            var coroutine = new CoroutineWrapper<object>(action, null, (exception) => Debug.LogError($"Exception on MainThread Queue: {exception.Message}"));
-            _executionQueue.Enqueue(() => { 
-                StartCoroutine(coroutine); 
-            });
-        }
-    }
-
-    /// <summary>
-    /// Locks the queue and adds the Action to the queue
-    /// </summary>
-    /// <param name="action">function that will be executed from the main thread.</param>
-    public static void Enqueue(Action action)
-    {
-        Instance().Enqueue(ActionWrapper(action));
-    }
-
-    /// <summary>
-    /// Locks the queue and adds the Action to the queue
-    /// </summary>
-    /// <param name="action">function that will be executed from the main thread.</param>
-    /// <param name="parameter">function parameter.</param>
-    public static void Enqueue<T>(Action<T> action, T parameter)
-    {
-        Instance().Enqueue(ActionWrapper(action, parameter));
-    }
-
-    /// <summary>
-    /// Locks the queue and adds the Action to the queue, returning a Task which is completed when the action completes
-    /// </summary>
-    /// <param name="action">function that will be executed from the main thread.</param>
-    /// <returns>A Task that can be awaited until the action completes</returns>
-    public Task EnqueueAsync(Action action)
-    {
-        var tcs = new TaskCompletionSource<bool>();
-
-        void WrappedAction()
-        {
-            try
-            {
-                action();
-                tcs.TrySetResult(true);
-            }
-            catch (Exception ex)
-            {
-                tcs.TrySetException(ex);
+                while (_executionQueue.Count > 0)
+                {
+                    _executionQueue.Dequeue().Invoke();
+                }
             }
         }
 
-        Enqueue(ActionWrapper(WrappedAction));
-        return tcs.Task;
-    }
-
-
-    private static IEnumerator ActionWrapper(Action a)
-    {
-        a();
-        yield return null;
-    }
-
-    private static IEnumerator ActionWrapper<T>(Action<T> function, T parameter)
-    {
-        yield return null;
-        function.Invoke(parameter);
-    }
-
-    private static UnityMainThreadDispatcher _instance;
-
-    private static bool Exists()
-    {
-        return _instance != null;
-    }
-
-    public static UnityMainThreadDispatcher Instance()
-    {
-        if (!Exists())
+        /// <summary>
+        /// Locks the queue and adds the IEnumerator to the queue
+        /// </summary>
+        /// <param name="action">IEnumerator function that will be executed from the main thread.</param>
+        public void Enqueue(IEnumerator action)
         {
-            throw new Exception(
-                "UnityMainThreadDispatcher could not find the UnityMainThreadDispatcher object. Please ensure you have added the MainThreadExecutor Prefab to your scene.");
+            lock (_executionQueue)
+            {
+                var coroutine = new CoroutineWrapper<object>(action, null, (exception) => Debug.LogError($"Exception on MainThread Queue: {exception.Message}"));
+                _executionQueue.Enqueue(() => { 
+                    StartCoroutine(coroutine); 
+                });
+            }
         }
 
-        return _instance;
-    }
-
-
-    void Awake()
-    {
-        if (_instance == null)
+        /// <summary>
+        /// Locks the queue and adds the Action to the queue
+        /// </summary>
+        /// <param name="action">function that will be executed from the main thread.</param>
+        public static void Enqueue(Action action)
         {
-            _instance = this;
-            DontDestroyOnLoad(this.gameObject);
+            Instance().Enqueue(ActionWrapper(action));
         }
-    }
 
-    void OnDestroy()
-    {
-        _instance = null;
+        /// <summary>
+        /// Locks the queue and adds the Action to the queue
+        /// </summary>
+        /// <param name="action">function that will be executed from the main thread.</param>
+        /// <param name="parameter">function parameter.</param>
+        public static void Enqueue<T>(Action<T> action, T parameter)
+        {
+            Instance().Enqueue(ActionWrapper(action, parameter));
+        }
+
+        /// <summary>
+        /// Locks the queue and adds the Action to the queue, returning a Task which is completed when the action completes
+        /// </summary>
+        /// <param name="action">function that will be executed from the main thread.</param>
+        /// <returns>A Task that can be awaited until the action completes</returns>
+        public Task EnqueueAsync(Action action)
+        {
+            var tcs = new TaskCompletionSource<bool>();
+
+            void WrappedAction()
+            {
+                try
+                {
+                    action();
+                    tcs.TrySetResult(true);
+                }
+                catch (Exception ex)
+                {
+                    tcs.TrySetException(ex);
+                }
+            }
+
+            Enqueue(ActionWrapper(WrappedAction));
+            return tcs.Task;
+        }
+
+
+        private static IEnumerator ActionWrapper(Action a)
+        {
+            a();
+            yield return null;
+        }
+
+        private static IEnumerator ActionWrapper<T>(Action<T> function, T parameter)
+        {
+            yield return null;
+            function.Invoke(parameter);
+        }
+
+        private static UnityMainThreadDispatcher _instance;
+
+        private static bool Exists()
+        {
+            return _instance != null;
+        }
+
+        public static UnityMainThreadDispatcher Instance()
+        {
+            if (!Exists())
+            {
+                throw new Exception(
+                    "UnityMainThreadDispatcher could not find the UnityMainThreadDispatcher object. Please ensure you have added the MainThreadExecutor Prefab to your scene.");
+            }
+
+            return _instance;
+        }
+
+
+        void Awake()
+        {
+            if (_instance == null)
+            {
+                _instance = this;
+                DontDestroyOnLoad(this.gameObject);
+            }
+        }
+
+        void OnDestroy()
+        {
+            _instance = null;
+        }
     }
 }
