+++ Runtime/Scripts/Tezos/API/Models/Alias.cs
@@ -1,6 +1,6 @@
 ﻿using System.Text.Json.Serialization;
 
-namespace Scripts.Tezos.API.Models
+namespace TezosSDK.Tezos.API.Models
 {
     public class Alias
     {
