+++ Runtime/Scripts/Tezos/TezosSingleton.cs
@@ -4,15 +4,15 @@ using System.Collections.Generic;
 using System.Text.Json;
 using Beacon.Sdk.Beacon.Permission;
 using Beacon.Sdk.Beacon.Sign;
-using Scripts.BeaconSDK;
-using Scripts.Helpers;
-using Scripts.Tezos.API;
-using Scripts.Tezos.API.Models.Filters;
-using Scripts.Tezos.API.Models.Tokens;
-using Scripts.Tezos.Wallet;
+using TezosSDK.Beacon;
+using TezosSDK.DesignPattern.Singleton;
+using TezosSDK.Helpers;
+using TezosSDK.Tezos.API;
+using TezosSDK.Tezos.API.Models.Filters;
+using TezosSDK.Tezos.API.Models.Tokens;
+using TezosSDK.Tezos.Wallet;
 
-
-namespace Scripts.Tezos
+namespace TezosSDK.Tezos
 {
     public class TezosSingleton : SingletonMonoBehaviour<TezosSingleton>, ITezosAPI
     {
