+++ Runtime/Scripts/Tezos/Wallet/WalletProviderType.cs
@@ -1,4 +1,5 @@
-﻿namespace Scripts.Tezos.Wallet
+﻿// ReSharper disable InconsistentNaming
+namespace TezosSDK.Tezos.Wallet
 {
     public enum WalletProviderType
     {
